Refactoring ChartPanel.java & Make ChartModel.java
